requirements.txt
# Python requirements for DMarket tools

# Main dependencies
python-telegram-bot>=20.0
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
pydantic>=2.0.0
aiofiles>=23.0.0
aiogram>=3.1.0

# Development dependencies
mypy>=1.0.0
ruff>=0.0.280
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0

# Type stubs for better type checking
types-pytest>=7.0.0
types-requests>=2.0.0
